Initial Application Version
The project has recently increased in scope; therefore, GitHub will be used for version control from this point forward. In the current state, the program uses a class object for anything effected by gravity/motion. As of now, forces do not exist, and movement is tied directly to velocity and acceleration, and the 1st object (Object 0) is given a highlight. 

Currently, 10 objects are created with random starting positions, starting velocities, sizes, and colors. No collisions occur between objects.  

Note that the velocity decays over time. This is meant to mimic inelastic collisions, but was only meant to be temporary, the next major step is to make motion dependent on forces and mass.